Refactored input and contact events
