+++ src/main/java/rx/libgdx/events/input/InputEvent.java
@@ -13,10 +13,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package rx.libgdx.events;
+package rx.libgdx.events.input;
 
-public class KeyDownEvent extends KeyUpOrDownEvent {
-    public KeyDownEvent(int keycode) {
-        super(keycode);
-    }
+public interface InputEvent {
+    // marker interface
 }
