+++ src/main/java/rx/libgdx/events/input/KeyDownEvent.java
@@ -13,16 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package rx.libgdx.events;
+package rx.libgdx.events.input;
 
-public class KeyTypedEvent implements InputEvent {
-    final char character;
-    
-    public KeyTypedEvent(char character) {
-        this.character = character;
-    }
-
-    public char getCharacter() {
-        return character;
+public class KeyDownEvent extends KeyUpOrDownEvent {
+    public KeyDownEvent(int keycode) {
+        super(keycode);
     }
 }
