+++ src/main/java/rx/libgdx/events/input/KeyTypedEvent.java
@@ -13,10 +13,16 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package rx.libgdx.events;
+package rx.libgdx.events.input;
 
-public class MouseMovedEvent extends ScreenCoordsEvent {
-    public MouseMovedEvent(int screenX, int screenY) {
-        super(screenX, screenY);
-    } 
+public class KeyTypedEvent implements InputEvent {
+    final char character;
+    
+    public KeyTypedEvent(char character) {
+        this.character = character;
+    }
+
+    public char getCharacter() {
+        return character;
+    }
 }
