+++ src/main/java/rx/libgdx/events/input/KeyUpEvent.java
@@ -13,17 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package rx.libgdx.events;
+package rx.libgdx.events.input;
 
-public abstract class TouchEvent extends ScreenCoordsEvent {
-    private final int pointer;
-
-    public TouchEvent(int screenX, int screenY, int pointer) {
-        super(screenX, screenY);
-        this.pointer = pointer;
-    }
-    
-    public int getPointer() {
-        return pointer;
+public class KeyUpEvent extends KeyUpOrDownEvent {
+    public KeyUpEvent(int keycode) {
+        super(keycode);
     }
 }
