+++ src/main/java/rx/libgdx/events/input/ScreenCoordsEvent.java
@@ -13,10 +13,22 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package rx.libgdx.events;
+package rx.libgdx.events.input;
 
-public class TouchDraggedEvent extends TouchEvent {
-  public TouchDraggedEvent(int screenX, int screenY, int pointer) {
-    super(screenX, screenY, pointer);
-  }
+abstract class ScreenCoordsEvent implements InputEvent {
+    private final int screenX;
+    private final int screenY;
+    
+    public ScreenCoordsEvent(int screenX, int screenY) {
+        this.screenX = screenX;
+        this.screenY = screenY;
+    }
+
+    public int getScreenX() {
+        return screenX;
+    }
+
+    public int getScreenY() {
+        return screenY;
+    }
 }
