+++ src/main/java/rx/libgdx/events/input/TouchDownEvent.java
@@ -13,8 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package rx.libgdx.events;
+package rx.libgdx.events.input;
 
-public interface InputEvent {
-    // marker interface
+public class TouchDownEvent extends TouchUpOrDownEvent {
+  public TouchDownEvent(int screenX, int screenY, int pointer, int button) {
+    super(screenX, screenY, pointer, button);
+  }
 }
