+++ src/main/java/rx/libgdx/events/input/TouchEvent.java
@@ -13,16 +13,17 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package rx.libgdx.events;
+package rx.libgdx.events.input;
 
-abstract class KeyUpOrDownEvent implements InputEvent {
-    private final int keycode;
-  
-    public KeyUpOrDownEvent(int keycode) {
-        this.keycode = keycode;
+public abstract class TouchEvent extends ScreenCoordsEvent {
+    private final int pointer;
+
+    public TouchEvent(int screenX, int screenY, int pointer) {
+        super(screenX, screenY);
+        this.pointer = pointer;
     }
-  
-    public int getKeycode() {
-        return keycode;
+    
+    public int getPointer() {
+        return pointer;
     }
 }
