+++ src/main/java/rx/libgdx/events/input/TouchUpOrDownEvent.java
@@ -13,22 +13,17 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package rx.libgdx.events;
+package rx.libgdx.events.input;
 
-abstract class ScreenCoordsEvent implements InputEvent {
-    private final int screenX;
-    private final int screenY;
+class TouchUpOrDownEvent extends TouchEvent {
+    private final int button;
     
-    public ScreenCoordsEvent(int screenX, int screenY) {
-        this.screenX = screenX;
-        this.screenY = screenY;
+    public TouchUpOrDownEvent(int screenX, int screenY, int pointer, int button) {
+        super(screenX, screenY, pointer);
+        this.button = button;
     }
 
-    public int getScreenX() {
-        return screenX;
-    }
-
-    public int getScreenY() {
-        return screenY;
+    public int getButton() {
+        return button;
     }
 }
